asyncifying index+database tests
